library api started

+++ models/wishlist.js
@@ -7,7 +7,7 @@ module.exports = function(sequelize, DataTypes) {
             allowNull: true
         },
         ISBN_13: {
-            type: DataTypes.INTEGER ,
+            type: DataTypes.INTEGER (13),
             allowNull: true
         },
         Title: {
